Feat[renderer]: move angle setup to RenderEnvUtils and rework it
