+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/utils/RenderEnvUtils.java
@@ -1,11 +1,14 @@
 package net.kdt.pojavlaunch.utils;
 
 import android.content.Context;
+import android.util.Log;
 
 import net.kdt.pojavlaunch.Architecture;
+import net.kdt.pojavlaunch.Tools;
 import net.kdt.pojavlaunch.plugins.LibraryPlugin;
 import net.kdt.pojavlaunch.prefs.LauncherPreferences;
 
+import java.io.File;
 import java.util.Map;
 
 /**
@@ -15,8 +18,12 @@ public class RenderEnvUtils {
 
     public static final String MESA_EGL = "libEGL_mesa.so";
     public static final String MESA_EGL_LEGACY = "libEGL_legacy.so";
+    public static final String ANGLE_EGL = "libEGL_angle.so";
+    public static final String ANGLE_OPENGL = "libGLESv2_angle.so";
+
 
     private static LibraryPlugin zink;
+    private static LibraryPlugin angle;
     /**
      * Setup environment for the mesa-based renderers. Does nothing if the renderer is not Mesa
      * @param context context
@@ -53,13 +60,48 @@ public class RenderEnvUtils {
     }
 
     /**
-     * Destroy zink plugin instance created during environment init if exists
+     * Setup ANGLE environment for LTW/GL4ES
+     * @param ctx Context
+     * @param envMap current environment map
+     * @return if setup succeeded
      */
-    public static void destroyZink(){
-        if(zink != null) {
-            zink = null;
-            System.gc();
+    public static boolean setupAngleEnv(Context ctx, Map<String, String> envMap) {
+        angle = LibraryPlugin.discoverPlugin(ctx, LibraryPlugin.ID_ANGLE_PLUGIN);
+        if (angle == null) return setupSystemAngle(envMap);
+        String[] angleLibs = {ANGLE_EGL, ANGLE_OPENGL};
+        if (!angle.checkLibraries(angleLibs)) {
+            Log.e("AngleEnvSetup", "AnglePlugin exists, but the ANGLE libraries are not present. Is the plugin corrupted?");
+            return setupSystemAngle(envMap);
         }
+        Log.i("JREUtils", "Using external ANGLE plugin");
+        envMap.put("LIBGL_EGL", angle.resolveAbsolutePath(angleLibs[0]));
+        envMap.put("LIBGL_GLES", angle.resolveAbsolutePath(angleLibs[1]));
+        return true;
+    }
+    private static boolean setupSystemAngle(Map<String, String> envMap){
+        File[] system = getSystemAngle();
+        if(system == null) return false;
+        Log.i("JREUtils", "Using system-provided ANGLE");
+        envMap.put("LIBGL_EGL", system[0].getAbsolutePath());
+        envMap.put("LIBGL_GLES", system[1].getAbsolutePath());
+        return true;
+    }
+
+    public static File[] getSystemAngle(){
+        File egl = new File(Architecture.is64BitsDevice() ? "/system/lib64" : "/system/lib", ANGLE_EGL);
+        File gles = new File(Architecture.is64BitsDevice() ? "/system/lib64" : "/system/lib", ANGLE_OPENGL);
+        if(egl.exists() && gles.exists())
+            return new File[]{egl, gles};
+        else return null;
+    }
+
+    /**
+     * Cleanup plugin instances created during environment init if exists
+     */
+    public static void cleanup(){
+        if(zink != null) zink = null;
+        if(angle != null) angle = null;
+        System.gc();
     }
 
     /**
@@ -82,5 +124,8 @@ public class RenderEnvUtils {
         if ((LauncherPreferences.PREF_ZINK_FORCE_LEGACY || GLInfoUtils.getGlInfo().isArm()) && zink != null)
             return zink.getLibraryPath();
         return null;
+    }
+    public static boolean hasAnglePlugin(){
+        return angle != null;
     }
 }
